requirements.txt
fastapi==0.123.9
nltk==3.9.1
pydantic==2.12.5
pymongo==4.6.0
pyparsing==2.4.7
python-decouple==3.8
rdflib==7.5.0
Requests==2.32.5
uvicorn==0.34.0